fix additional APIENTRY warnings
