+++ common/calibration-model.cpp
@@ -1,9 +1,10 @@
-#include <glad/glad.h>
 #include "calibration-model.h"
 #include "model-views.h"
 #include "os.h"
+#include "ux-window.h"
 
 #include "../src/ds5/ds5-private.h"
+
 
 using namespace rs2;
