+++ common/calibration-model.h
@@ -1,7 +1,9 @@
+// License: Apache 2.0. See LICENSE file in root directory.
+// Copyright(c) 2021 Intel Corporation. All Rights Reserved.
+
 #pragma once
 
 #include <librealsense2/rs.hpp>
-#include "ux-window.h"
 
 namespace librealsense
 {
@@ -10,6 +12,8 @@ namespace librealsense
 
 namespace rs2
 {
+    class ux_window;
+
     class calibration_model
     {
     public:
