+++ common/measurement.h
@@ -2,7 +2,6 @@
 // Copyright(c) 2020 Intel Corporation. All Rights Reserved.
 #pragma once
 
-#include "rendering.h"
 #include "model-views.h"
 
 #include <vector>
